# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: EDP Control Plane API
  description: |
    Enterprise data control plane for registering sources and datasets,
    authoring and deploying pipelines, enforcing data quality and governance
    policy, tracking lineage and impact, orchestrating runs, and recording an
    append-only audit trail.

    ## Conventions

    * All collection endpoints are paginated with `limit` and `offset`, and
      sorted with `sort_by` and `sort_dir`. See the shared parameters.
    * Errors always use the same envelope: `error.code`, `error.message`,
      `error.details` and a top-level `request_id` that matches the
      `X-Request-Id` response header.
    * Mutations are `PATCH` rather than `PUT`, so a caller updates only the
      fields it means to change.
    * Request bodies are validated before any state is written; a validation
      failure returns `422` with one entry per offending field.
  version: 0.1.0
  contact:
    name: Data Platform Engineering
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://edp.internal/api/v1
    description: Internal deployment
security:
  - bearerAuth: []
...
paths:
...
  /quality-rules:
    get:
      tags: [quality]
      operationId: listQualityRules
      summary: List quality rules
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: search
          in: query
          schema: { type: string }
        - name: dataset_id
          in: query
          schema: { type: string, format: uuid }
        - name: pipeline_id
          in: query
          schema: { type: string, format: uuid }
        - name: rule_type
          in: query
          schema: { type: array, items: { type: string } }
        - name: severity
          in: query
          schema: { type: array, items: { type: string, enum: [info, warning, error, critical] } }
        - name: dimension
          in: query
          description: Repeatable; the dimension field is multi-valued.
          schema: { type: array, items: { type: string } }
        - name: enabled
          in: query
          schema: { type: boolean }
        - name: blocking
          in: query
          schema: { type: boolean }
        - name: tag
          in: query
          schema: { type: array, items: { type: string } }
      responses:
        '200':
          description: A page of quality rules.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QualityRulePage' }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [quality]
      operationId: createQualityRule
      summary: Create a quality rule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QualityRuleWrite' }
      responses:
        '201':
          description: The created rule.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QualityRule' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /quality-rules/{ruleID}:
    parameters:
      - $ref: '#/components/parameters/RuleID'
    get:
      tags: [quality]
      operationId: getQualityRule
      summary: Fetch a quality rule
      responses:
        '200':
          description: The rule.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QualityRule' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [quality]
      operationId: updateQualityRule
      summary: Update a quality rule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QualityRuleWritePartial' }
      responses:
        '200':
          description: The updated rule.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QualityRule' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      tags: [quality]
      operationId: deleteQualityRule
      summary: Delete a quality rule
      responses:
        '204': { description: Deleted. }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /quality-checks:
    get:
      tags: [quality]
      operationId: listQualityChecks
      summary: List check runs across all rules
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: rule_id
          in: query
          schema: { type: string, format: uuid }
        - name: dataset_id
          in: query
          schema: { type: string, format: uuid }
        - name: pipeline_run_id
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: array, items: { type: string } }
      responses:
        '200':
          description: A page of check runs.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QualityCheckPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /quality-checks/{checkID}:
    parameters:
      - name: checkID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [quality]
      operationId: getQualityCheck
      summary: Fetch a check run
      responses:
        '200':
          description: The check run, including observed and expected values.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QualityCheck' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A JWT verified against the configured JWKS. Issuer, audience and expiry are
        all checked. When authentication is disabled the API runs without any
        middleware and every request is attributed to a development actor.
  parameters:
    Limit:
      name: limit
      in: query
      description: Page size.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      description: Rows to skip.
      schema: { type: integer, minimum: 0, default: 0 }
    SortBy:
      name: sort_by
      in: query
      description: Column to sort on. Unknown columns are rejected.
      schema: { type: string }
    SortDir:
      name: sort_dir
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
...
    RuleID:
      name: ruleID
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: The query string or path parameters were malformed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationFailed:
      description: The body failed validation. Every offending field is listed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code.
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - METHOD_NOT_ALLOWED
                - CONFLICT
                - VALIDATION_FAILED
                - RATE_LIMITED
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
                - NOT_IMPLEMENTED
            message:
              type: string
              description: Human-readable summary. Safe to show a user.
            details:
              type: array
              description: Present for validation failures.
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
        request_id:
          type: string
          format: uuid
          description: Matches the X-Request-Id header, for correlating with logs.
    PageMeta:
      type: object
      required: [limit, offset, total]
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
...
    QualityRule:
      type: object
      properties:
        id: { type: string, format: uuid }
        dataset_id: { type: string, format: uuid, nullable: true }
        pipeline_id: { type: string, format: uuid, nullable: true }
        name: { type: string }
        description: { type: string }
        rule_type: { type: string }
        dimension: { type: array, items: { type: string } }
        severity: { type: string, enum: [info, warning, error, critical] }
        expression: { type: string }
        threshold: { type: number }
        blocking:
          type: boolean
          description: When true a failure fails the pipeline run rather than only warning.
        enabled: { type: boolean }
        tags: { type: array, items: { type: string } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    QualityRuleWrite:
      type: object
      required: [name, rule_type, expression]
      properties:
        dataset_id: { type: string, format: uuid }
        pipeline_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        rule_type: { type: string }
        dimension: { type: array, items: { type: string } }
        severity: { type: string, enum: [info, warning, error, critical] }
        expression: { type: string }
        threshold: { type: number }
        blocking: { type: boolean }
        enabled: { type: boolean }
        tags: { type: array, items: { type: string } }
    QualityRuleWritePartial:
      allOf:
        - $ref: '#/components/schemas/QualityRuleWrite'
    QualityCheck:
      type: object
      properties:
        id: { type: string, format: uuid }
        rule_id: { type: string, format: uuid }
        dataset_id: { type: string, format: uuid, nullable: true }
        pipeline_run_id: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [passed, failed, warning, error, skipped] }
        observed_value: { type: number, nullable: true }
        expected_value: { type: number, nullable: true }
        row_count: { type: integer, nullable: true }
        message: { type: string }
        checked_at: { type: string, format: date-time }
...
    QualityRulePage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/QualityRule' } }
    QualityCheckPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/QualityCheck' } }
...
    Page:
      type: object
      description: Every list endpoint returns the same envelope.
      required: [items, meta]
      properties:
        items:
          type: array
          items: {}
        meta: { $ref: '#/components/schemas/PageMeta' }